Fixed fatal design flaw in Queue.
The way I had designed Queue originally was with an 'await null inside while true' approach, which actually was creating endless promises and being very inefficient. It didn't matter before but now, trying to run problem 11 I finally had a long-running enough program that I actually ran out of heap memory.

What I really needed was to be able to return a promise (which I was already doing), but have a way to resolve that promise from outside of itself (elsewhere within the Queue class). To do this, I created a class called Deferred which owns a promise, and the resolve/reject functions that are given to the promise.

Now Queue's pop function can return the promise immediately, and Queue's push function can be responsible for resolving the promise. This makes way more sense, since Queue.push is actually the first time we'll know that a new value is available and therefore the promise can be resolved.